Refatored code for Atom feed format writer in service so that we can support the Json format implementation.
